requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
neo4j==5.14.1
torch==2.1.0
torch-geometric==2.4.0
pydantic==2.5.0
pydantic-settings==2.1.0
apscheduler==3.10.4
httpx==0.25.2